Tweaked some things regarding my game's physics

+++ game.h
@@ -15,8 +15,8 @@ private:
     Graphics graphics;
 
     // timing
-    double dt;
+    float dt;
     Uint64 performance_frequency;
     Uint64 prev_counter;
-    double lag;
+    float lag;
 };
